[#18] feat :rocket: : 컴포넌트 계층 구조로 파일 디렉토리 변경
- 불필요한 기본 CRA 파일들 삭제
